reorganize /docs directories for En and Fr content
En and Fr content are divided into two main folder in /docs: /fr and /en.
A landing page (/collections-model/) is added where users can select their preferred language.
The language button is now located to the right of the top menu.
A modal with disclaimer of english-only content is activate when clicking on the language button from any pages of TM and SPS.
In teh French top menu, Modèle cible and Spécification des chemins sémantiques are linked directly the their English version. Thinking of creating a page for each with the Disclaimer and tell user to visit the English version.